move some cluster/server info functions into source files and add some access ops
